Bugfix: IP checking when in IP mode isnt the same as with FQDNs...
git-svn-id: file:///home/code/svn/ovd/trunk@357 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/servers.php
@@ -163,6 +163,12 @@ if (isset($_REQUEST['action']) && $_REQUEST['action'] == 'web_port' && isset($_R
 }
 
 if (isset($_REQUEST['action']) && $_REQUEST['action'] == 'delete' && isset($_REQUEST['fqdn'])) {
+	$sessions = Sessions::getByServer($_REQUEST['fqdn']);
+	if (count($sessions) > 0) {
+		popup_error(_('Unable to delete a server when there are active sessions on it!'));
+		redirect($_SERVER['HTTP_REFERER']);
+	}
+
 	Abstract_Server::delete($_REQUEST['fqdn']);
 
 	redirect('servers.php?action=list');
